requirements.txt
PyYAML>=6.0
torch>=1.11.0+cu113
torchvision>=0.12.0+cu113
opencv-python>=4.6.0.66
numpy>=1.22.4
transformers>=4.19.4
jaconv>=0.3
Pillow>=9.1.1
scikit-image>=0.19.2
scipy>=1.8.1